Rename components folder to snippets for clarity
