+++ app/shared/snippets/avatar.tsx
