+++ app/shared/snippets/checkbox.tsx
