+++ app/shared/snippets/close-button.tsx
