+++ app/shared/snippets/slider.tsx
